regionliveness: add support for querying unavailable timestamps
To add support for being able to clamp SQL session renewals to the the
lifespan of a region, the region liveness prober interface needs to be
updated to support querying timestamps. This patch updates the region
liveness prober to support querying unavailable_at times add the ability
to retrieve timestamps.

Release note: None