sqlinstance: make initial scan for range feed liveness aware
Previously, the initial scan used by the SQL instance range feed to
monitor changes was not region-liveness aware. As a result, if a region
were to have gone down, we could hang waiting for a reply from a region.
This patch makes the initial scan region aware and probes regions before
querying them.

Release note: None