roachprod: destroy MIG resources
After adding MIG support, the way resources are destroyed for `managed` clusters
need to be updated. Instead of deleting the instances, as was done previously,
we destroy the MIG that will destroy the VMs as well. Only after the MIGs have
been destroyed we can also delete the instance template originally created for
the cluster.

Resolves: #118419
Epic: CRDB-33832

Release Note: None