sql: add telemetry for configured and used isolation level
This telemetry will allow us to see when non-SERIALIZABLE isolation
levels are used, and also when they are requested but automatically
upgraded to SERIALIZABLE due to settings not being enabled yet.

This does not track instances of SERIALIZABLE transactions, since that
is the default, so the noise would overwhelm the other telemetry data.

The telemetry is under the following feature counters:
- sql.txn.isolation.executed_at.level
- sql.txn.isolation.upgraded_from.level

Release note: None