roachtest: log when tests use spot VMs
It's currently not easy to determine which tests used spot VMs in a
nightly run. This commit updates the test runner so that, when we
decide to run a test with spot VMs, we log that decision. This should
make it trivial to check which  tests ran on spot VMs in any given run
and compare that to the failures observed to get a sense of the
failure ratio.

Epic: none

Release note: None