build: alphabetize go runtime patches
We currently maintain a set of patches to the Go runtime. To-date, our
various patches have accumulated in a single patch file. Ideally, our
set of patches should be periodically re-applied to the base of the Go
runtime that we build, to ensure that our patches still apply cleanly.
However, the lack of ordering in the patch currently file makes this
difficult.

Alphabetize the patches to avoid future diff cheese.

The changes in this patch were generated using the following:

```
_patch_file="$CRDB_SRC_ROOT/build/teamcity/internal/release/build-and-publish-patched-go/diff.patch"
$ cd go
$ git checkout go1.21.5
$ git apply "$_patch_file"
$ git diff > "$_patch_file"
```

Release note: None.