roachtest: increase `version-upgrade` timeout in remote runs
With recent changes in the `mixedversion` framework, we could be
waiting for background workloads a little longer than usual in remote
runs, which could cause timeouts in that setting.

To avoid that, we increase the test's timeout in remote runs to 2
hours, keeping the existing 30 minute timeout in local (CI) runs. This
is safe because the framework itself reduces wait time when run in
local mode.

Fixes: #118619

Release note: None