roachtest: reduce post restore workload runtime
Previously during an OR roachtest, after the online restore download job
finished, the workload would run until the test nearly timed out. Given the 400
GB roachtest timeout was 5 hours, the roachtest workload would run for about
4.5 hours. This is wasteful and doesn't add much coverage. This patch reduces
this test's timeout to 1 hour and configures all OR roachtests to run the post
OR worklaod for a maximum of 10 minutes.

Epic: none

Release note: none