catalog/lease: optimize counting leases by including region
Previously, the plans for scanning the leasing table would always do
full table scans because the region was never constrained.  This was
problematic especially when a large number of leases existed in single
region environments. To address this, this patch optimizes the leasing
count queries for the single region scenario.

Epic: none

Release note: None