sql: disable streamer in fewer cases in internal executor
This commit makes it so that the streamer is disabled in the internal
executor in exactly the same scenario when DistSQL was disabled in the
previous commit. This change polishes the fix in
ed3f640510fa2d993f9cf4508cd0f8b9c53733d5 to be more precise. It's in
a separate commit since it won't be backported (unlike the previous
one).

Release note: None