roachtest: don't use spot VMs in weekly test runs
Many of the weekly tests take several hours or days to finish. It's
wasteful to run a test for several hours only to give up due to
preemption. Since the running time is so long, it's also unlikely
these tests will ever run to completion.

Epic: none

Release note: None