restore: re-run downloads until poller is done
Previously the download pass would only run once, since the
API contract of Download() is that when it completes there are
not external files in its span.

However if somehow there is still an external file in the span,
for example if it was added while Download() was running,
then the poller may keep observing non-zero extenral files and
keep the job running, even though there is no longer a download
task running that will change that, causing it to hang forever.

This changes the download task to keep running repeatedly as long as
the poller is observing that there is something to download, so that
if there are still external files after a pass completes for the
reasons described above, they will be downloaded on in a subsequent pass.

Release note: none.
Epic: none.