workflows: build helper binaries before running tests
Up until this point we were trying to use `bazel shutdown` to get the
`bazel` server to come down when GitHub cancels a running build, so we
can run a helper binary build after the fact. I have had a lot of
difficulty getting this to work properly and reliably. Instead, I
re-order things so that we perform the build of the helper binary first.
Then, we have no dependency on Bazel after a cancellation comes in.

Epic: CRDB-8308
Release note: None