changefeedccl: add nil checking for avroDataRecord.refreshTypeMetadata
Previously, the "after" field was populated across all envelope formats.
However, in commit c00bf0f, row data was moved to the “after” field for envelope
= wrapped and to “record” field for other envelope formats. But the commit did
not add a nil check for registered.schema.after which is only populated for
wrapped envelope. registered.schema.after is now nil for other envelope
formats, and calling refreshTypeMetadata on it could lead to node panics. In
addition, this commit did not add the logic calling refreshTypeMetadata on the
registered.schema.record for other envelope formats. This patch addresses the
issue by adding a defensive nil check before invoking refreshTypeMetadata and
adding refreshTypeMetadata for the record field when it's set to match the
original intents of the commit.

Fixes: #119428
Release note: None