kvserver: add allocator integration skeleton
This is likely logically a bit flawed, but the core intent remains.

We need to register non-mma operations, get some changeIDs and then
enact them.

In the future, there should be a simulation and non-simulation execution
component, which uses the `AllocatorSync` to apply state updates. The
reason for the separation, is the different in execution model. The
production code blocks on its changes, while the simulation code has a
non-blocking model.

Epic: none
Release note: None