backup: add warning when no replica moved from AdminScatter
Previously, we introduced a new field NoReplicasMoved to
AdminScatterResponse. This is to provide more visibility into
the scatter operation to indicate whether AdminScatter actually
successfully moved replicas.

This commit just logs a warning if NoReplicasMoved is true
from the DR side. Note that retrying from the client side is
not implemented, so no behaviour changes exist yet. This remains
as an ongoing issue tracked in #144856.

Informs: #144856
Part of: #144560
Release note: none