mma: skip maxFractionPendingIncrease if reported load is the same as adjusted load
Previously, MMA set ss.maxFractionPendingIncrease and
ss.maxFractionPendingDecrease to 1000 when the reported load was 0. This could
occur when some dimension (e.g. write bandwidth during read-only workloads).
This patch skips the adjustment when both reported and adjusted loads are
equal for a dimension. Note that if the reported load is
zero but the adjusted load is non-zero, we may still fall back to 1000.