mma: fix bug in sortTargetCandidateSetAndPick
The intention of the code was to be structured around sets representing
equivalence classes, such that a later set is only considered if none
of the earlier sets had a member that was discarded and had pending
changes. Prior to this change, this criteria was arbitrarily applied in
the middle of a set. So if two stores {s1, s2} were in the same set and
s1 was discarded and had pending changes we would not consider s2.
Now we will include s2 and stop when the next set starts.