kvserver: register external changes with mma
Previously, allocator sync was called before and after applying changes, but
only to update the store pool. This commit extends it to also register changes
with the mma allocator. If mma is enabled at registration time, allocator sync
tracks the change and informs mma during PostApply. If mma is disabled, it
continues to only update the store pool.