mmaintegration: integrate allocator sync to mma store rebalancer
Previously, mma PreApply was introduced to register mma changes with allocator
sync. This commit integrates it into the mma store rebalancer. To handle
failures that occur before a change is even registered (such as when a replica
doesn’t exist), this commit also adds, MarkChangesAsFailed, in allocator sync.
This function bypasses the trackedChanges map since the change is not meant to
be applied. It must only be called on failure, and allocator sync is responsible
for notifying MMA of the failure.